# Kubernetes-Native-Microservice-WithQuarkusAndMicroProfile/Part02-Developing-Microservices/02-Nicer-Response-Exception/account-service/src/main/java/ir/farhad/model/entities/AccountEntity.java
package ir.farhad.model.entities;

import ir.farhad.model.enums.AccountStatus;

import java.math.BigDecimal;

public class AccountEntity {
    public Long accountNumber;
    public Long customerNumber;
    public String customerName;
    public BigDecimal balance;
    public AccountStatus accountStatus = AccountStatus.OPEN;
    public AccountEntity() {
    }
    public AccountEntity(Long accountNumber, Long customerNumber, String
            customerName, BigDecimal balance) {
        this.accountNumber = accountNumber;
        this.customerNumber = customerNumber;
        this.customerName = customerName;
        this.balance = balance;
    }
    public void markOverdrawn() {
        accountStatus = AccountStatus.OVERDRAWN;
    }

    public void removeOverdrawnStatus() {
        accountStatus = AccountStatus.OPEN;
    }
    public void close() {
        accountStatus = AccountStatus.CLOSED;
        balance = BigDecimal.valueOf(0);
    }
    public void withdrawFunds(BigDecimal amount) {
        balance = balance.subtract(amount);
    }
    public void addFunds(BigDecimal amount) {
        balance = balance.add(amount);
    }
    public BigDecimal getBalance() {
        return balance;
    }
    public Long getAccountNumber() {
        return accountNumber;
    }
    public String getCustomerName() {
        return customerName;
    }
    public AccountStatus getAccountStatus() {
        return accountStatus;
    }

}